no longer install -pthread variants
upstream's use of more modern features of libgpg-error means that we
do not need specific libgpgme-pthread variants any more.

(see upstream commit 09b64554328445e99a8cc78fc34ea49c2ea2e7f9 for more
explanation)

this is likely to cause some weird upgrade paths for packages that
previously depended on libgpgme-pthread.so.11 -- we need to think
clarify what to do about this.

Since some previous packages might be dependent on
libgpgme-pthread.so.11, though, we also ship a link to it so that
those packages still work.