dropping C++ and QT symbols files
C++ and QT symbols files appear to be dynamic and change depending on
the architecture.  This results in needing to do multiple round-trips
through the buildds to learn how the different exported symbols vary.
yuck!

For simplicity's sake for now, we'll just go with a tighter dependency
chain (dh_makeshlibs -V) and not bother with trying to track them
explicitly.